+++ solver/src/test/java/org/chocosolver/solver/search/loop/LNSTest.java
@@ -9,6 +9,10 @@
  */
 package org.chocosolver.solver.search.loop;
 
+import static java.lang.Math.ceil;
+import static org.chocosolver.solver.search.strategy.Search.domOverWDegSearch;
+import static org.chocosolver.solver.search.strategy.Search.lastConflict;
+
 import org.chocosolver.solver.Cause;
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.Solution;
@@ -16,7 +20,11 @@ import org.chocosolver.solver.Solver;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.search.limits.BacktrackCounter;
 import org.chocosolver.solver.search.loop.lns.INeighborFactory;
-import org.chocosolver.solver.search.loop.lns.neighbors.*;
+import org.chocosolver.solver.search.loop.lns.neighbors.INeighbor;
+import org.chocosolver.solver.search.loop.lns.neighbors.PropagationGuidedNeighborhood;
+import org.chocosolver.solver.search.loop.lns.neighbors.RandomNeighborhood;
+import org.chocosolver.solver.search.loop.lns.neighbors.ReversePropagationGuidedNeighborhood;
+import org.chocosolver.solver.search.loop.lns.neighbors.SequenceNeighborhood;
 import org.chocosolver.solver.search.loop.move.Move;
 import org.chocosolver.solver.search.loop.move.MoveBinaryDFS;
 import org.chocosolver.solver.search.loop.move.MoveLNS;
@@ -31,10 +39,6 @@ import org.chocosolver.util.tools.ArrayUtils;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-
-import static java.lang.Math.ceil;
-import static org.chocosolver.solver.search.strategy.Search.domOverWDegSearch;
-import static org.chocosolver.solver.search.strategy.Search.lastConflict;
 
 /**
  * <br/>
@@ -164,6 +168,52 @@ public class LNSTest {
         // find the new best solution
         int bw = 0, bp = 0;
         while (model.getSolver().solve()) {
+            bp = power.getValue();
+            bw = scalar.getValue();
+        }
+        r.printShortStatistics();
+        Assert.assertEquals(bp, 6937);
+        Assert.assertEquals(bw, 1092);
+    }
+
+    @Test(groups="1s", timeOut=60000)
+    public void testLoadSolution() {
+        // First, the model: here a simple knapsack pb ...
+        int[] capacities = {99, 1101};
+        int[] volumes = {54, 12, 47, 33, 30, 65, 56, 57, 91, 88, 77, 99, 29, 23, 39, 86, 12, 85, 22, 64};
+        int[] energies = {38, 57, 69, 90, 79, 89, 28, 70, 38, 71, 46, 41, 49, 43, 36, 68, 92, 33, 84, 90};
+
+        Model model = new Model();
+        int nos = 20;
+        // occurrence of each item
+        IntVar[] objects = new IntVar[nos];
+        for (int i = 0; i < nos; i++) {
+            objects[i] = model.intVar("o_" + (i + 1), 0, (int) ceil(capacities[1] / volumes[i]), true);
+        }
+        final IntVar power = model.intVar("power", 0, 99999, true);
+        IntVar scalar = model.intVar("weight", capacities[0], capacities[1], true);
+        model.scalar(objects, volumes, "=", scalar).post();
+        model.scalar(objects, energies, "=", power).post();
+        model.knapsack(objects, scalar, power, volumes, energies).post();
+        model.setObjective(Model.MAXIMIZE, power);
+        // ... end of modelling
+
+        Solver r = model.getSolver();
+        Solution sol = r.findSolution();
+        r.printShortStatistics();
+        // let's start the reparation of the previous solution
+        // so, reset the search
+        r.reset();
+        // declaring a neighborhood for LNS, here a random one
+        RandomNeighborhood rnd = new RandomNeighborhood(objects, 1, 0);
+        // initialize it with the previous (best) solution
+        // declare LNS
+        r.setLNS(rnd, sol);
+        // limit search
+        r.limitNode(3000);
+        // find the new best solution
+        int bw = 0, bp = 0;
+        while (r.solve()) {
             bp = power.getValue();
             bw = scalar.getValue();
         }