+++ solver/CHANGES.md
@@ -11,11 +11,12 @@ Multi-modules and JPMS-ready.
 ### Major features:
 - Move `cutoffseq`, `choco-sat`, `choco-solver`, `pf4cs`, `choco-parsers` and `samples` projects into a (maven) multi-modules project, JPMS-ready 
 - Default AC algorithm for `AllDifferent` is now from IJCAI-18 "A Fast Algorithm for Generalized Arc Consistency of the Alldifferent Constraint", Zhang et al. (#644)
+- LNS can now be defined with a solution as bootstrap.
 
 ### Deprecated API (to be removed in next release):
 
 ### Closed issues and pull requests:
-\#617, #633, #639, #647
+\#617, #633, #639, #646, #647
 
 4.10.2 - 14 Oct 2019
 -------------------